+++ pkg/httpx/range.go
@@ -3,6 +3,7 @@ package httpx
 import (
 	"errors"
 	"fmt"
+	"net/http"
 	"strconv"
 	"strings"
 )
@@ -22,31 +23,35 @@ func (rng Range) Size() int64 {
 	return rng.End - rng.Start + 1
 }
 
-func ParseRangeHeader(h string, size int64) (Range, error) {
+func ParseRangeHeader(header http.Header, size int64) (*Range, error) {
+	h := header.Get(HeaderRange)
+	if h == "" {
+		return nil, nil
+	}
 	if size <= 0 {
-		return Range{}, fmt.Errorf("size %d cannot be equal or less than zero", size)
+		return nil, fmt.Errorf("size %d cannot be equal or less than zero", size)
 	}
 	rangeUnitPrefix := RangeUnit + "="
 	if !strings.HasPrefix(h, rangeUnitPrefix) {
-		return Range{}, errors.New("invalid range unit")
+		return nil, errors.New("invalid range unit")
 	}
 	ranges := strings.Split(strings.TrimPrefix(h, rangeUnitPrefix), ",")
 	if len(ranges) != 1 {
-		return Range{}, errors.New("multiple ranges not supported")
+		return nil, errors.New("multiple ranges not supported")
 	}
 	parts := strings.SplitN(strings.TrimSpace(ranges[0]), "-", 2)
 	if len(parts) != 2 {
-		return Range{}, errors.New("invalid range format")
+		return nil, errors.New("invalid range format")
 	}
 
 	// Suffix byte range
 	if parts[0] == "" {
 		suffixLen, err := strconv.ParseInt(parts[1], 10, 64)
 		if err != nil {
-			return Range{}, err
+			return nil, err
 		}
 		if suffixLen <= 0 {
-			return Range{}, fmt.Errorf("invalid suffix range %d", suffixLen)
+			return nil, fmt.Errorf("invalid suffix range %d", suffixLen)
 		}
 		if suffixLen > size {
 			suffixLen = size
@@ -55,16 +60,16 @@ func ParseRangeHeader(h string, size int64) (Range, error) {
 			Start: size - suffixLen,
 			End:   size - 1,
 		}
-		return rng, nil
+		return &rng, nil
 	}
 
 	rng := Range{}
 	start, err := strconv.ParseInt(parts[0], 10, 64)
 	if err != nil {
-		return Range{}, err
+		return nil, err
 	}
 	if start < 0 {
-		return Range{}, fmt.Errorf("invalid start %d", start)
+		return nil, fmt.Errorf("invalid start %d", start)
 	}
 	rng.Start = start
 	if parts[1] == "" {
@@ -72,7 +77,7 @@ func ParseRangeHeader(h string, size int64) (Range, error) {
 	} else {
 		end, err := strconv.ParseInt(parts[1], 10, 64)
 		if err != nil {
-			return Range{}, err
+			return nil, err
 		}
 		if end >= size {
 			end = size - 1
@@ -80,9 +85,9 @@ func ParseRangeHeader(h string, size int64) (Range, error) {
 		rng.End = end
 	}
 	if rng.Start > rng.End {
-		return Range{}, fmt.Errorf("start %d cannot be larger than end %d", rng.Start, rng.End)
+		return nil, fmt.Errorf("start %d cannot be larger than end %d", rng.Start, rng.End)
 	}
-	return rng, nil
+	return &rng, nil
 }
 
 type ContentRange struct {